Extract the content handlers.
